+++ tests/testthat/test-get_outputs.R
@@ -1,0 +1,8 @@
+testthat::test_that("", {
+  q <- qenv()
+  q1 <- eval_code(q, expression(iris, mtcars))
+  testthat::expect_identical(
+    get_outputs(q1),
+    list(iris, mtcars)
+  )
+})