+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.code 0.6.1.9003
+# teal.code 0.6.1.9004
 
 ### Enhancements